Move emails functions connected to account to account module
