+++ saleor/account/emails.py
@@ -95,3 +95,34 @@ def _send_delete_confirmation_email(recipient_email, delete_url):
         context=ctx,
         **send_kwargs,
     )
+
+
+def send_set_password_email_with_url(redirect_url, user, staff=False):
+    """Trigger sending a set password email for the given customer/staff."""
+    template_type = "staff" if staff else "customer"
+    template = f"dashboard/{template_type}/set_password"
+    token = default_token_generator.make_token(user)
+    _send_set_user_password_email_with_url.delay(
+        user.email, redirect_url, token, template
+    )
+
+
+@app.task
+def _send_set_user_password_email_with_url(
+    recipient_email, redirect_url, token, template_name
+):
+    params = urlencode({"email": recipient_email, "token": token})
+    password_set_url = urlsplit(redirect_url)
+    password_set_url = password_set_url._replace(query=params)
+    _send_set_password_email(recipient_email, password_set_url.geturl(), template_name)
+
+
+def _send_set_password_email(recipient_email, password_set_url, template_name):
+    send_kwargs, ctx = get_email_context()
+    ctx["password_set_url"] = password_set_url
+    send_templated_mail(
+        template_name=template_name,
+        recipient_list=[recipient_email],
+        context=ctx,
+        **send_kwargs,
+    )
